final 2016 mac checkin

+++ pylab_util.py
@@ -34,9 +34,9 @@ def _build_pdf_path(pathin):
     return pdfpath
 
 
-def mysave(path_in, fi=None, ext='.eps', dpi=100, pdfcrop=1):
+def mysave(path_in, fi=None, ext='.eps', dpi=100, pdfcrop=1, kwargs={}):
     myfig = _get_fig(fi)
-    mplutil.mysave(path_in, myfig, ext=ext, dpi=dpi)
+    mplutil.mysave(path_in, myfig, ext=ext, dpi=dpi,kwargs=kwargs)
     if pdfcrop:
         pdfpath = _build_pdf_path(path_in)
         if os.path.exists(pdfpath):
@@ -45,9 +45,9 @@ def mysave(path_in, fi=None, ext='.eps', dpi=100, pdfcrop=1):
             os.system(cmd)
 
 def save_and_crop_png(path_in, fi=None, dpi=450, pad_and_resize=True, \
-                      clean=True, width=1600, height=1000):
+                      clean=True, width=1600, height=1000, kwargs={'bbox_inches':'tight'}):
     myfig = _get_fig(fi)
-    mplutil.mysave(path_in, myfig, ext='.png', dpi=dpi)
+    mplutil.mysave(path_in, myfig, ext='.png', dpi=dpi, kwargs=kwargs)
     cmd = 'mytrim.py %s' % path_in
     os.system(cmd)
     if pad_and_resize:
@@ -57,7 +57,8 @@ def save_and_crop_png(path_in, fi=None, dpi=450, pad_and_resize=True, \
                (fn2, width, height)
         os.system(cmd2)
         if clean:
-            os.remove(fn2)
+            if os.path.exists(fn2):
+                os.remove(fn2)
     if clean:
         os.remove(path_in)
